Improve usability on large files.
With larger brif files, it could take a wee while for the GUI to
show anything.

Add a thread to do the file loading/processing. This allows the GUI
to come up straight away while the file is still being processed.

NOTE: Most of the processing is in the prettyfying of the display;
fonts, colours etc and not the actual file loading.