Fix the displaying of negative values.
In add_dp() we check if we got a value < 100 in order to prepend
it with a 0. If we got a minus value, say -12400 then that would
have a 0 prepended and become 0-12400.

Fix the check for < 100 to make it more strict that the value is
actually between 0 and 100 both +ve and -ve.