Fix a segfault introduced in 44ab3a6a
In 44ab3a6a we set the window title to include the name of the
currently viewed brif file. However if we opened the file
selection and just cancelled it, we'd end up trying to set the
window title to null.

Only set the window title if it's not null.