tools: Add a tool to create a csv of values from a brif file
bv2csv takes the given brif file and generates a CSV of the values
contained within.

It generates one line per Purchasing Card Item. For multiple lines
within the same invoice, it displays the Main Record and Purchasing
Card details only on the first line.

Usage is simply:

    $ bv2csv /path/to/brif_file

it will generate a brif_file.csv in the current working directory.

Signed-off-by: Andrew Clayton <andrew@pccl.info>