viewbrif: Fix a memory leak in cb_file_chooser()
The filename variable that is set by gtk_file_chooser_get_filename()
must be free'd.

Re-work cb_file_chooser() slightly so that we can free this before
calling gtk_widget_destroy() which needs to be done last.

We also need to make a copy of the filename string passed into
read_file_thread() as filename is being free'd before it's finished with
it.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>