tools/bv2csv: Simplify add_dp()
add_dp() takes a value such as 1750 and returns 17.50, i.e return the
given number to the specified number of decimal places.

Previously this function was doing all kinds of string manipulation.
Really all we need to do is divide the number by 10^dp. Where dp is the
required number of decimal places.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>