viewbrif:  More fix's to loading new files
When loading a new file into viewbrif, GTK was producing the following
warnings

    Gtk-CRITICAL **: gtk_text_buffer_emit_insert: assertion
    `g_utf8_validate (text, len, NULL)' failed

in read_file() and then sometimes the mmap() would fail and the
filename that was loaded would not be displayed.

This seems to be issues with threading and data being clobbered.

I seem to have fixed (hidden?) this issue by reducing the amount of
gdk_threads_enter()/gdk_threads_leave() pairs in read_file().

I have only removed some pairs from the bit which is only done once per
file and doesn't cause any noticable increase in blocking of the UI.

There are however likely to be other such related problems lurking.

Signed-off-by: Andrew Clayton <andrew@pccl.info>