viewbrif: Re-work threading
Even after the previous attempts at fixing the threading issues, some
still remained, e.g trying to load a new file while one is currently in
progress.

The general consensus seems to be that you should limit GTK function
calls to the main thread. We do this by using g_idle_add{_full} in the
file reading thread to have the functions that actually do the GTK stuff
be called from the main thread.

read_file() is called from the main thread and sets up the text views,
it then creates a thread, file_read_thread() which does the mmap and
then sets the appropriate functions to be called from the main thread
as mentioned above.

This allows us to get rid of all the gdk_threads_{enter,leave}
function calls as well as the gdk_flush calls.

This should make things much more robust and seems to also make the
interface a little more responsive.

We can also now load a file while one is currently in progress. To do
this we track the id's as returned by the calls to g_idle_add{_full}
and stick them in an array. We can then use g_source_remove to remove
any pending functions still to be run and then load up the new file.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>