viewbrif: Fix some potential undefined behaviour
In process_line() we used sprintf()'s to set some variables. We then did
further sprintf()'s to re-set those variables using the variable as both
the source and destination, e.g

	sprintf(myvar, "(%s)", myvar);

However this can lead to undefined behaviour depending on things like
compiler and libc versions.

This has been working on GCC and GLibc as shipped with the past many
Fedora releases, however, best not to rely on this behaviour.

We can actually just get rid of the second lot of sprintf()'s and do
what we need to in the first lot.

This problem was flagged by the cppcheck(1)[0] static analyser.

[0] - http://cppcheck.sourceforge.net/

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>