viewbrif: Fix a buffer overflow in do_purchasing_card_item()
When compiling with -D_FORTIFY_SOURCE=2 -fstack-protector viewbrif was
immediately crashing upon loading a brif file with the following
backtrace

    *** buffer overflow detected ***: ./viewbrif terminated
    ======= Backtrace: =========
    /lib64/libc.so.6(+0x77a8d)[0x7f443c711a8d]
    /lib64/libc.so.6(__fortify_fail+0x37)[0x7f443c7ae667]
    /lib64/libc.so.6(+0x1127f0)[0x7f443c7ac7f0]
    /lib64/libc.so.6(+0x111d59)[0x7f443c7abd59]
    /lib64/libc.so.6(_IO_default_xsputn+0x80)[0x7f443c715550]
    /lib64/libc.so.6(_IO_vfprintf+0x396e)[0x7f443c6e6f0e]
    /lib64/libc.so.6(__vsprintf_chk+0x8c)[0x7f443c7abdec]
    /lib64/libc.so.6(__sprintf_chk+0x7d)[0x7f443c7abd3d]
.   /viewbrif[0x404a83]
    /lib64/libglib-2.0.so.0(g_main_context_dispatch+0x15a)[0x7f443d1aea8a]
    /lib64/libglib-2.0.so.0(+0x49e20)[0x7f443d1aee20]
    /lib64/libglib-2.0.so.0(g_main_loop_run+0xc2)[0x7f443d1af142]
    /lib64/libgtk-x11-2.0.so.0(gtk_main+0xb7)[0x7f443ec58f37]
.   /viewbrif[0x403688]
    /lib64/libc.so.6(__libc_start_main+0xf0)[0x7f443c6ba700]
    /viewbrif[0x4036b9]

Examining the core file under GDB showed this to be comming from the
sprintf() call in do_purchasing_card_item() where we were overflowing
the buffer that holds the line header string, e.g

    Line no: 649, Purchasing Card Item

This buffer was only 35 chars long which wouldn't be enough to hold the
above line plus the '\n' and '\0'.

Make this buffer 64 chars long and use snprintf() to set it.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>