viewbrif: Fix buffer overlows in display_stats()
When compiling with -Wp,-D_FORTIFY_SOURCE=2 -fstack-protector viewbrif
was immedaitely crashing while to trying to load a large brif file with
the following backtrace

    *** buffer overflow detected ***: ./viewbrif terminated
    ======= Backtrace: =========
    /lib64/libc.so.6(+0x77a8d)[0x7f01a3be4a8d]
    /lib64/libc.so.6(__fortify_fail+0x37)[0x7f01a3c81667]
    /lib64/libc.so.6(+0x1127f0)[0x7f01a3c7f7f0]
    /lib64/libc.so.6(+0x111d59)[0x7f01a3c7ed59]
    /lib64/libc.so.6(_IO_default_xsputn+0x80)[0x7f01a3be8550]
    /lib64/libc.so.6(_IO_vfprintf+0x396e)[0x7f01a3bb9f0e]
    /lib64/libc.so.6(__vsprintf_chk+0x8c)[0x7f01a3c7edec]
    /lib64/libc.so.6(__sprintf_chk+0x7d)[0x7f01a3c7ed3d]
    ./viewbrif[0x403fe7]
    /lib64/libglib-2.0.so.0(g_main_context_dispatch+0x15a)[0x7f01a4681a8a]
    /lib64/libglib-2.0.so.0(+0x49e20)[0x7f01a4681e20]
    /lib64/libglib-2.0.so.0(g_main_loop_run+0xc2)[0x7f01a4682142]
    /lib64/libgtk-x11-2.0.so.0(gtk_main+0xb7)[0x7f01a612bf37]
    ./viewbrif[0x4036d8]
    /lib64/libc.so.6(__libc_start_main+0xf0)[0x7f01a3b8d700]
    ./viewbrif[0x403709]

Examining the resultant core file under GDB showed this was the
sprintf() in display_stats().

Looking at it it is pretty clear it's totaly broken. I was mallocing a
variable using things like sizeof(brif_stats.file_size), which in this
case is a off_t, to get the size, this is of course complete nonsense.

There isn't really any need to use malloc() as opposed to just a small
static buffer.

So remove all the mallocs/free's (we also don't need to do any memset's)
and use use a small 32 char buffer instead, also use snprintf() instead
of sprintf() for extra safety.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>